+++ core/components/romanesco/elements/snippets/06_formulas/f_resource/resourcemediapath.snippet.php
@@ -10,7 +10,8 @@
  * have its own images that page 456 could not reference.
  *
  * USAGE
- * [[resourceMediaPath? &pathTpl=`assets/businesses/{id}/` &createFolder=`1`]]
+ * [[resourceMediaPath? &pathTpl=`assets/businesses/{id}/`]]
+ * [[resourceMediaPath? &pathTpl=`assets/resourceimages/{id}/` &checkTVs=`mymigxtv`]]
  * [[resourceMediaPath? &pathTpl=`assets/test/{breadcrumb}`]]
  * [[resourceMediaPath? &pathTpl=`assets/test/{breadcrumb}` &breadcrumbdepth=`2`]]
  *
@@ -20,6 +21,7 @@
  *		Available placeholders: {id}, {pagetitle}, {parent}
  * &docid (optional) integer page id
  * &createFolder (optional) boolean whether to create folder or not
+ * &checkTVs (optional) comma-separated list of TVs to check, before directory is created
  *
  * @var modX $modx
  * @var array $scriptProperties
@@ -29,9 +31,12 @@ $pathTpl = $modx->getOption('pathTpl', $scriptProperties, '');
 $docid = $modx->getOption('docid', $scriptProperties, '');
 $createfolder = $modx->getOption('createFolder', $scriptProperties, false);
 $tvname = $modx->getOption('tvname', $scriptProperties, '');
+$checktvs = $modx->getOption('checkTVs', $scriptProperties, false);
 
 $path = '';
+$fullpath = '';
 $createpath = false;
+$fallbackpath = $modx->getOption('fallbackPath', $scriptProperties, 'assets/migxfallback/');
 
 if (empty($pathTpl)) {
     $modx->log(modX::LOG_LEVEL_ERROR, '[resourceMediaPath]: pathTpl not specified.');
@@ -55,31 +60,45 @@ if (empty($docid)) {
     if (is_object($modx->resource)) {
         $docid = $modx->resource->get('id');
     }
-    //on backend
+    //on manager resource/update page
     else {
         $createpath = $createfolder;
-        // We do this to read the &id param from an Ajax request
-        $parsedUrl = parse_url($_SERVER['HTTP_REFERER']);
-        parse_str($parsedUrl['query'], $parsedQuery);
 
-        if (isset($parsedQuery['amp;id'])) {
-            $docid = (int)$parsedQuery['amp;id'];
-        } elseif (isset($parsedQuery['id'])) {
-            $docid = (int)$parsedQuery['id'];
+        // Read the &id param from an Ajax request
+        $parsedUrl = parse_url($_SERVER['HTTP_REFERER']);
+        if ($parsedUrl['query'] ?? '') {
+            parse_str($parsedUrl['query'], $parsedQuery);
+        }
+
+        // Avoid docid to be set to parent container
+        $requestAction = $_REQUEST['a'] ?? '';
+        $action = $parsedQuery['a'] ?? '';
+        if (!$action && $requestAction || $action == $requestAction) {
+            $docid = $_REQUEST['id'] ?? '';
+        }
+        elseif ($action === 'resource/update') {
+            $docid = (int)$parsedQuery['amp;id'] ?? (int)$parsedQuery['id'] ?? 0;
         }
     }
 }
 
 if (empty($docid)) {
-    $modx->log(modX::LOG_LEVEL_ERROR, '[resourceMediaPath]: docid could not be determined.');
-    return;
+    $modx->log(modX::LOG_LEVEL_DEBUG, '[resourceMediaPath]: docid could not be determined.');
 }
 
-if ($resource = $modx->getObject('modResource', $docid)) {
+if (empty($docid) || empty($pathTpl)) {
+    $path = $fallbackpath;
+    $fullpath = $modx->getOption('base_path') . $fallbackpath;
+    $createpath = true;
+}
+
+if (empty($fullpath) && $resource = $modx->getObject('modResource', $docid)) {
     $path = $pathTpl;
     $ultimateParent = '';
     if (strstr($path, '{breadcrumb}') || strstr($path, '{ultimateparent}')) {
-        $parentids = $modx->getParentIds($docid);
+        $depth = $modx->getOption('breadcrumbdepth', $scriptProperties, 10);
+        $ctx = $resource->get('context_key');
+        $parentids = $modx->getParentIds($docid, $depth, array('context' => $ctx));
         $breadcrumbdepth = $modx->getOption('breadcrumbdepth', $scriptProperties, count($parentids));
         $breadcrumbdepth = $breadcrumbdepth > count($parentids) ? count($parentids) : $breadcrumbdepth;
         if (count($parentids) > 1) {
@@ -115,22 +134,36 @@ if ($resource = $modx->getObject('modResource', $docid)) {
     }
     if ($user = $modx->user) {
         $path = str_replace('{username}', $modx->user->get('username'), $path);
+        $path = str_replace('{userid}', $modx->user->get('id'), $path);
     }
 
     $fullpath = $modx->getOption('base_path') . $path;
 
-    if ($createpath && !file_exists($fullpath)) {
-
-        $permissions = octdec('0' . (int)($modx->getOption('new_folder_permissions', null, '755', true)));
-        if (!@mkdir($fullpath, $permissions, true)) {
-            $modx->log(modX::LOG_LEVEL_ERROR, sprintf('[resourceMediaPath]: could not create directory %s).', $fullpath));
-        } else {
-            chmod($fullpath, $permissions);
+    if ($createpath && $checktvs){
+        $createpath = false;
+        if ($template) {
+            $tvs = explode(',',$checktvs);
+            foreach ($tvs as $tv){
+                if ($template->hasTemplateVar($tv)){
+                    $createpath = true;
+                }
+            }
         }
+
     }
 
-    return $path;
 } else {
-    $modx->log(modX::LOG_LEVEL_ERROR, sprintf('[resourceMediaPath]: resource not found (page id %s).', $docid));
-    return;
+    $modx->log(modX::LOG_LEVEL_DEBUG, sprintf('[resourceMediaPath]: resource not found (page id %s).', $docid));
 }
+
+if ($createpath && !file_exists($fullpath)) {
+
+    $permissions = octdec('0' . (int)($modx->getOption('new_folder_permissions', null, '755', true)));
+    if (!@mkdir($fullpath, $permissions, true)) {
+        $modx->log(modX::LOG_LEVEL_DEBUG, sprintf('[resourceMediaPath]: could not create directory %s).', $fullpath));
+    } else {
+        chmod($fullpath, $permissions);
+    }
+}
+
+return $path;